Got a jump working properly in a test scene
Yep, jumping works if given a time to the apex and a height. It bothers
me that it's a little bit inaccurate on the timing, but it's a start.
I'd like to make the final jump method more customizable, ie: time to
apex, hang time, time from apex to ground, height, etc. I'll look into
making the jump options feel good.