Updated Jump Mechanics
The test script for jumping mechanics now allows you to define a jump by
its overall height as well as the distance it covers laterally (at a
constant movement speed) before and after the apex. This allows the jump
to have a floaty rise and a sharp fall, or vice-versa. Next up is
variable jump heights and a toggle-able double jump option.