Implemented Variable Jump Height
You can now define a minimum and maximum height for the jump. Next up is
double jumping, which may or may not have variable heights applied to it
as well. In the future, I may make the variable jump height based on how
long you hold the button down instead of the position in the jump where
the jump key is released.