Broke it a little...
I tried to start including initial position in the gravity calculations,
but the minimum jump height gravity equation is off for some reason. Not
sure why, but this needs to work in order to make double jumping
functional, as well as during the actual game when the player can be on
any height platform. Really need to think this one through...