Fixed my mistake
Turns out that when calculating gravity for early releases of the jump
button, both chunks of code are typically called (below min height
between min and max height) because the velocity carries the player
between min and max heights ever so slightly even when we're aiming for
minimum height. The error was actually in the section where we set
gravity between min and max heights. The line was "gravity =
-(Mathf.Pow(initialVelocity, 2.0f)) / (2.0f * (transform.position.y -
initialPositionY) - minJumpHeight); ", but it shouldn't have been
subtracting minJumpHeight. It's been fixed now.