Placed gravity stuff into a new method
I think it's probably best to keep the long chain of gravity logic in
its own method. Seems oay so far, but it can definitely be cleaned up
and made easier to read. I also started working more on the double jump
stuff. It's working sometimes, but for some reason the height is off by
a bit every once in a while (every couple times I run it). Definitely
need to sort that out.