Cleaned up jumping, got double jump working, and started on the raycast manager
The jumping gravity logic is now neat and tidy and in its own method.
Double jumping is working well, in part because of the newly cleaned up
logic. Raycast manager has been started (just decided to start it from
scratch). There's some code that I pulled from the old raycast manager
that I don't quite understand (it's the origin setting equation for
raycasting) but I'll take some more time and get my head wrapped around
it later. For now, it has basic functionality. I'll work on it more once
the planning for it is complete.